Split PlanViewerControl.Properties.cs into feature partials
Properties.cs was a 1772-line catch-all whose name said "Properties" but
which actually built six different right-side panels. Peel the non-core
panels into partials of the same class (pure mechanical move, no behavior
change):

- PlanViewerControl.MissingIndexes.cs — ShowMissingIndexes
- PlanViewerControl.Parameters.cs — ShowParameters + param-cell/annotation
  builders + unresolved-variable detection
- PlanViewerControl.WaitStats.cs — ShowWaitStats + wait categorization
- PlanViewerControl.RuntimeSummary.cs — ShowRuntimeSummary + ShowServerContext

Properties.cs keeps the General/Object/etc. ShowPropertiesPanel core plus
the shared AddPropertySection/AddPropertyRow builders and close handlers,
dropping from 1772 to 1091 lines. (The per-node time-attribution helpers
already moved to Core's NodeTimeAttribution earlier in this branch.)
