+++ src/PlanViewer.App/Controls/PlanViewerControl.Properties.cs
@@ -1082,691 +1082,10 @@ public partial class PlanViewerControl : UserControl
         }
     }
 
-    private void ShowMissingIndexes(List<MissingIndex> indexes)
-    {
-        MissingIndexContent.Children.Clear();
-
-        if (indexes.Count > 0)
-        {
-            // Update expander header with count
-            MissingIndexHeader.Text = $"  Missing Index Suggestions ({indexes.Count})";
-
-            // Build each missing index row manually (no ItemsControl template binding)
-            foreach (var mi in indexes)
-            {
-                var itemPanel = new StackPanel { Margin = new Thickness(0, 4, 0, 0) };
-
-                var headerRow = new StackPanel { Orientation = Orientation.Horizontal };
-                headerRow.Children.Add(new TextBlock
-                {
-                    Text = mi.Table,
-                    FontWeight = FontWeight.SemiBold,
-                    Foreground = new SolidColorBrush(Color.Parse("#E4E6EB")),
-                    FontSize = 12
-                });
-                headerRow.Children.Add(new TextBlock
-                {
-                    Text = $" \u2014 Impact: ",
-                    Foreground = new SolidColorBrush(Color.Parse("#E4E6EB")),
-                    FontSize = 12
-                });
-                headerRow.Children.Add(new TextBlock
-                {
-                    Text = $"{mi.Impact:F1}%",
-                    Foreground = new SolidColorBrush(Color.Parse("#FFB347")),
-                    FontSize = 12
-                });
-                itemPanel.Children.Add(headerRow);
-
-                if (!string.IsNullOrEmpty(mi.CreateStatement))
-                {
-                    itemPanel.Children.Add(new SelectableTextBlock
-                    {
-                        Text = mi.CreateStatement,
-                        FontFamily = new FontFamily("Consolas"),
-                        FontSize = 11,
-                        Foreground = TooltipFgBrush,
-                        TextWrapping = TextWrapping.Wrap,
-                        Margin = new Thickness(12, 2, 0, 0)
-                    });
-                }
-
-                MissingIndexContent.Children.Add(itemPanel);
-            }
-
-            MissingIndexEmpty.IsVisible = false;
-        }
-        else
-        {
-            MissingIndexHeader.Text = "Missing Index Suggestions";
-            MissingIndexEmpty.IsVisible = true;
-        }
-    }
-
-    private void ShowParameters(PlanStatement statement)
-    {
-        ParametersContent.Children.Clear();
-        ParametersEmpty.IsVisible = false;
-
-        var parameters = statement.Parameters;
-
-        if (parameters.Count == 0)
-        {
-            var localVars = FindUnresolvedVariables(statement.StatementText, parameters, statement.RootNode);
-            if (localVars.Count > 0)
-            {
-                ParametersHeader.Text = "Parameters";
-                AddParameterAnnotation(
-                    $"Local variables detected ({string.Join(", ", localVars)}) — values not captured in plan XML",
-                    "#FFB347");
-            }
-            else
-            {
-                ParametersHeader.Text = "Parameters";
-                ParametersEmpty.IsVisible = true;
-            }
-            return;
-        }
-
-        ParametersHeader.Text = $"Parameters ({parameters.Count})";
-
-        var allCompiledNull = parameters.All(p => p.CompiledValue == null);
-        var hasCompiled = parameters.Any(p => p.CompiledValue != null);
-        var hasRuntime = parameters.Any(p => p.RuntimeValue != null);
-
-        // Build a 4-column grid: Name | Data Type | Compiled | Runtime
-        // Only show Compiled/Runtime columns if at least one param has that value
-        var colDef = "Auto,Auto"; // Name, DataType always shown
-        int compiledCol = -1, runtimeCol = -1;
-        int nextCol = 2;
-        if (hasCompiled)
-        {
-            colDef += ",*";
-            compiledCol = nextCol++;
-        }
-        if (hasRuntime)
-        {
-            colDef += ",*";
-            runtimeCol = nextCol++;
-        }
-        // If neither compiled nor runtime, still add one value column for "?"
-        if (!hasCompiled && !hasRuntime)
-        {
-            colDef += ",*";
-            compiledCol = nextCol++;
-        }
-
-        var grid = new Grid { ColumnDefinitions = new ColumnDefinitions(colDef) };
-        int rowIndex = 0;
-
-        // Header row
-        grid.RowDefinitions.Add(new RowDefinition(GridLength.Auto));
-        AddParamCell(grid, rowIndex, 0, "Parameter", "#7BCF7B", FontWeight.SemiBold);
-        AddParamCell(grid, rowIndex, 1, "Data Type", "#7BCF7B", FontWeight.SemiBold);
-        if (compiledCol >= 0)
-            AddParamCell(grid, rowIndex, compiledCol, hasCompiled ? "Compiled" : "Value", "#7BCF7B", FontWeight.SemiBold);
-        if (runtimeCol >= 0)
-            AddParamCell(grid, rowIndex, runtimeCol, "Runtime", "#7BCF7B", FontWeight.SemiBold);
-        rowIndex++;
-
-        foreach (var param in parameters)
-        {
-            grid.RowDefinitions.Add(new RowDefinition(GridLength.Auto));
-
-            // Name
-            AddParamCell(grid, rowIndex, 0, param.Name, "#E4E6EB", FontWeight.SemiBold);
-
-            // Data type
-            AddParamCell(grid, rowIndex, 1, param.DataType, "#E4E6EB");
-
-            // Compiled value
-            if (compiledCol >= 0)
-            {
-                var compiledText = param.CompiledValue ?? (allCompiledNull ? "" : "?");
-                var compiledColor = param.CompiledValue != null ? "#E4E6EB"
-                    : allCompiledNull ? "#E4E6EB" : "#E57373";
-                AddParamCell(grid, rowIndex, compiledCol, compiledText, compiledColor);
-            }
-
-            // Runtime value — amber if it differs from compiled
-            if (runtimeCol >= 0)
-            {
-                var runtimeText = param.RuntimeValue ?? "";
-                var sniffed = param.RuntimeValue != null
-                    && param.CompiledValue != null
-                    && param.RuntimeValue != param.CompiledValue;
-                var runtimeColor = sniffed ? "#FFB347" : "#E4E6EB";
-                var tooltip = sniffed
-                    ? "Runtime value differs from compiled — possible parameter sniffing"
-                    : null;
-                AddParamCell(grid, rowIndex, runtimeCol, runtimeText, runtimeColor, tooltip: tooltip);
-            }
-
-            rowIndex++;
-        }
-
-        ParametersContent.Children.Add(grid);
-
-        // Annotations
-        if (allCompiledNull && parameters.Count > 0)
-        {
-            var hasOptimizeForUnknown = statement.StatementText
-                .Contains("OPTIMIZE", StringComparison.OrdinalIgnoreCase)
-                && Regex.IsMatch(statement.StatementText, @"OPTIMIZE\s+FOR\s+UNKNOWN", RegexOptions.IgnoreCase);
-
-            if (hasOptimizeForUnknown)
-            {
-                AddParameterAnnotation(
-                    "OPTIMIZE FOR UNKNOWN — optimizer used average density estimates instead of sniffed values",
-                    "#6BB5FF");
-            }
-            else
-            {
-                AddParameterAnnotation(
-                    "OPTION(RECOMPILE) — parameter values embedded as literals, not sniffed",
-                    "#FFB347");
-            }
-        }
-
-        var unresolved = FindUnresolvedVariables(statement.StatementText, parameters, statement.RootNode);
-        if (unresolved.Count > 0)
-        {
-            AddParameterAnnotation(
-                $"Unresolved variables: {string.Join(", ", unresolved)} — not in parameter list",
-                "#FFB347");
-        }
-    }
-
-    private static void AddParamCell(Grid grid, int row, int col, string text, string color,
-        FontWeight fontWeight = default, string? tooltip = null)
-    {
-        var tb = new TextBlock
-        {
-            Text = text,
-            FontSize = 11,
-            FontWeight = fontWeight == default ? FontWeight.Normal : fontWeight,
-            Foreground = new SolidColorBrush(Color.Parse(color)),
-            Margin = new Thickness(0, 2, 10, 2),
-            TextTrimming = TextTrimming.CharacterEllipsis,
-            MaxWidth = 200
-        };
-        // Name and DataType columns are short — no need for max width
-        if (col <= 1)
-            tb.MaxWidth = double.PositiveInfinity;
-        if (tooltip != null)
-            ToolTip.SetTip(tb, tooltip);
-        else if (text.Length > 30)
-            ToolTip.SetTip(tb, text);
-        Grid.SetRow(tb, row);
-        Grid.SetColumn(tb, col);
-        grid.Children.Add(tb);
-    }
-
-    private void AddParameterAnnotation(string text, string color)
-    {
-        ParametersContent.Children.Add(new TextBlock
-        {
-            Text = text,
-            FontSize = 11,
-            FontStyle = FontStyle.Italic,
-            Foreground = new SolidColorBrush(Color.Parse(color)),
-            TextWrapping = TextWrapping.Wrap,
-            Margin = new Thickness(0, 6, 0, 0)
-        });
-    }
-
-    private static List<string> FindUnresolvedVariables(string queryText, List<PlanParameter> parameters,
-        PlanNode? rootNode = null)
-    {
-        var unresolved = new List<string>();
-        if (string.IsNullOrEmpty(queryText))
-            return unresolved;
-
-        var extractedNames = new HashSet<string>(
-            parameters.Select(p => p.Name), StringComparer.OrdinalIgnoreCase);
-
-        // Collect table variable names from the plan tree so we don't misreport them as local variables
-        var tableVarNames = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
-        if (rootNode != null)
-            CollectTableVariableNames(rootNode, tableVarNames);
-
-        var matches = Regex.Matches(queryText, @"@\w+", RegexOptions.IgnoreCase);
-        var seenVars = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
-
-        foreach (Match match in matches)
-        {
-            var varName = match.Value;
-            if (seenVars.Contains(varName) || extractedNames.Contains(varName))
-                continue;
-            if (varName.StartsWith("@@", StringComparison.OrdinalIgnoreCase))
-                continue;
-            if (tableVarNames.Contains(varName))
-                continue;
-
-            seenVars.Add(varName);
-            unresolved.Add(varName);
-        }
-
-        return unresolved;
-    }
-
-    private static void CollectTableVariableNames(PlanNode node, HashSet<string> names)
-    {
-        if (!string.IsNullOrEmpty(node.ObjectName) && node.ObjectName.StartsWith("@"))
-        {
-            // ObjectName is like "@t.c" — extract the table variable name "@t"
-            var dotIdx = node.ObjectName.IndexOf('.');
-            var tvName = dotIdx > 0 ? node.ObjectName[..dotIdx] : node.ObjectName;
-            names.Add(tvName);
-        }
-        foreach (var child in node.Children)
-            CollectTableVariableNames(child, names);
-    }
-
-    private void ShowWaitStats(List<WaitStatInfo> waits, List<WaitBenefit> benefits, bool isActualPlan)
-    {
-        WaitStatsContent.Children.Clear();
-
-        if (waits.Count == 0)
-        {
-            WaitStatsHeader.Text = "Wait Stats";
-            WaitStatsEmpty.Text = isActualPlan
-                ? "No wait stats recorded"
-                : "No wait stats (estimated plan)";
-            WaitStatsEmpty.IsVisible = true;
-            return;
-        }
-
-        WaitStatsEmpty.IsVisible = false;
-
-        // Build benefit lookup
-        var benefitLookup = new Dictionary<string, double>(StringComparer.OrdinalIgnoreCase);
-        foreach (var wb in benefits)
-            benefitLookup[wb.WaitType] = wb.MaxBenefitPercent;
-
-        var sorted = waits.OrderByDescending(w => w.WaitTimeMs).ToList();
-        var maxWait = sorted[0].WaitTimeMs;
-        var totalWait = sorted.Sum(w => w.WaitTimeMs);
-
-        // Update expander header with total
-        WaitStatsHeader.Text = $"  Wait Stats \u2014 {totalWait:N0}ms total";
-
-        // Build a single Grid for all rows so columns align
-        // Name, bar, duration, and benefit columns
-        var grid = new Grid
-        {
-            ColumnDefinitions = new ColumnDefinitions("Auto,*,Auto,Auto")
-        };
-        for (int i = 0; i < sorted.Count; i++)
-            grid.RowDefinitions.Add(new RowDefinition(GridLength.Auto));
-
-        for (int i = 0; i < sorted.Count; i++)
-        {
-            var w = sorted[i];
-            var barFraction = maxWait > 0 ? (double)w.WaitTimeMs / maxWait : 0;
-            var color = GetWaitCategoryColor(GetWaitCategory(w.WaitType));
-
-            // Wait type name — colored by category
-            var nameText = new TextBlock
-            {
-                Text = w.WaitType,
-                FontSize = 12,
-                Foreground = new SolidColorBrush(Color.Parse(color)),
-                VerticalAlignment = VerticalAlignment.Center,
-                Margin = new Thickness(0, 2, 10, 2)
-            };
-            Grid.SetRow(nameText, i);
-            Grid.SetColumn(nameText, 0);
-            grid.Children.Add(nameText);
-
-            // Bar — semi-transparent category color, compact proportional indicator
-            var barColor = Color.Parse(color);
-            var colorBar = new Border
-            {
-                Width = Math.Max(4, barFraction * 60),
-                Height = 14,
-                Background = new SolidColorBrush(Color.FromArgb(0x60, barColor.R, barColor.G, barColor.B)),
-                CornerRadius = new CornerRadius(2),
-                HorizontalAlignment = HorizontalAlignment.Left,
-                VerticalAlignment = VerticalAlignment.Center,
-                Margin = new Thickness(0, 2, 8, 2)
-            };
-            Grid.SetRow(colorBar, i);
-            Grid.SetColumn(colorBar, 1);
-            grid.Children.Add(colorBar);
-
-            // Duration text
-            var durationText = new TextBlock
-            {
-                Text = $"{w.WaitTimeMs:N0}ms ({w.WaitCount:N0} waits)",
-                FontSize = 12,
-                Foreground = new SolidColorBrush(Color.Parse("#E4E6EB")),
-                VerticalAlignment = VerticalAlignment.Center,
-                Margin = new Thickness(0, 2, 8, 2)
-            };
-            Grid.SetRow(durationText, i);
-            Grid.SetColumn(durationText, 2);
-            grid.Children.Add(durationText);
-
-            // Benefit % (if available)
-            if (benefitLookup.TryGetValue(w.WaitType, out var benefitPct) && benefitPct > 0)
-            {
-                var benefitText = new TextBlock
-                {
-                    Text = $"up to {benefitPct:N0}%",
-                    FontSize = 11,
-                    Foreground = new SolidColorBrush(Color.Parse("#8b949e")),
-                    VerticalAlignment = VerticalAlignment.Center,
-                    Margin = new Thickness(0, 2, 0, 2)
-                };
-                Grid.SetRow(benefitText, i);
-                Grid.SetColumn(benefitText, 3);
-                grid.Children.Add(benefitText);
-            }
-        }
-
-        WaitStatsContent.Children.Add(grid);
-
-    }
-
-    private void ShowRuntimeSummary(PlanStatement statement)
-    {
-        RuntimeSummaryContent.Children.Clear();
-
-        var labelColor = "#E4E6EB";
-        var valueColor = "#E4E6EB";
-
-        var grid = new Grid
-        {
-            ColumnDefinitions = new ColumnDefinitions("Auto,*")
-        };
-        int rowIndex = 0;
-
-        void AddRow(string label, string value, string? color = null)
-        {
-            grid.RowDefinitions.Add(new RowDefinition(GridLength.Auto));
-
-            var labelText = new TextBlock
-            {
-                Text = label,
-                FontSize = 11,
-                Foreground = new SolidColorBrush(Color.Parse(labelColor)),
-                HorizontalAlignment = HorizontalAlignment.Left,
-                Margin = new Thickness(0, 1, 8, 1)
-            };
-            Grid.SetRow(labelText, rowIndex);
-            Grid.SetColumn(labelText, 0);
-            grid.Children.Add(labelText);
-
-            var valueText = new TextBlock
-            {
-                Text = value,
-                FontSize = 11,
-                Foreground = new SolidColorBrush(Color.Parse(color ?? valueColor)),
-                Margin = new Thickness(0, 1, 0, 1)
-            };
-            Grid.SetRow(valueText, rowIndex);
-            Grid.SetColumn(valueText, 1);
-            grid.Children.Add(valueText);
-
-            rowIndex++;
-        }
-
-        // Efficiency thresholds: white >= 40%, orange >= 20%, red < 20%.
-        // Loosened per Joe's feedback (#215 C1): for memory grants, moderate
-        // utilization (e.g. 60%) is fine — operators can spill near their max,
-        // so we shouldn't flag anything above a real over-grant threshold.
-        static string EfficiencyColor(double pct) => pct >= 40 ? "#E4E6EB"
-            : pct >= 20 ? "#FFB347" : "#E57373";
-
-        // Memory grant color tiers (#215 C1 + E8 + E9): over-used grant (red),
-        // any operator spilled (orange), otherwise tier by utilization.
-        static string MemoryGrantColor(double pctUsed, bool hasSpill)
-        {
-            if (pctUsed > 100) return "#E57373";
-            if (hasSpill) return "#FFB347";
-            if (pctUsed >= 40) return "#E4E6EB";
-            if (pctUsed >= 20) return "#FFB347";
-            return "#E57373";
-        }
-
-        // E7: rename the panel title for estimated plans
-        var isEstimated = statement.QueryTimeStats == null;
-        RuntimeSummaryTitle.Text = isEstimated ? "Predicted Runtime" : "Runtime Summary";
-
-        var hasSpillInTree = statement.RootNode != null && HasSpillInPlanTree(statement.RootNode);
-
-        // E11: order — Elapsed → CPU:Elapsed → DOP → CPU → Compile → Memory → Used → Optimization → CE Model → Cost.
-        // Extra Avalonia-only rows (threads, UDF, cached plan size) kept near their logical neighbors.
-
-        if (statement.QueryTimeStats != null)
-        {
-            AddRow("Elapsed", $"{statement.QueryTimeStats.ElapsedTimeMs:N0}ms");
-            if (statement.QueryTimeStats.ElapsedTimeMs > 0)
-            {
-                long externalWaitMs = 0;
-                foreach (var w in statement.WaitStats)
-                    if (BenefitScorer.IsExternalWait(w.WaitType))
-                        externalWaitMs += w.WaitTimeMs;
-                var effectiveCpu = Math.Max(0L, statement.QueryTimeStats.CpuTimeMs - externalWaitMs);
-                var ratio = (double)effectiveCpu / statement.QueryTimeStats.ElapsedTimeMs;
-                AddRow("CPU:Elapsed", ratio.ToString("N2"));
-            }
-        }
-
-        // DOP + parallelism efficiency
-        if (statement.DegreeOfParallelism > 0)
-        {
-            var dopText = statement.DegreeOfParallelism.ToString();
-            string? dopColor = null;
-            if (statement.QueryTimeStats != null &&
-                statement.QueryTimeStats.ElapsedTimeMs > 0 &&
-                statement.QueryTimeStats.CpuTimeMs > 0 &&
-                statement.DegreeOfParallelism > 1)
-            {
-                long externalWaitMs = 0;
-                foreach (var w in statement.WaitStats)
-                    if (BenefitScorer.IsExternalWait(w.WaitType))
-                        externalWaitMs += w.WaitTimeMs;
-                var effectiveCpu = Math.Max(0, statement.QueryTimeStats.CpuTimeMs - externalWaitMs);
-                var speedup = (double)effectiveCpu / statement.QueryTimeStats.ElapsedTimeMs;
-                var efficiency = Math.Min(100.0, (speedup - 1.0) / (statement.DegreeOfParallelism - 1.0) * 100.0);
-                efficiency = Math.Max(0.0, efficiency);
-                dopText += $" ({efficiency:N0}% efficient)";
-                dopColor = EfficiencyColor(efficiency);
-            }
-            AddRow("DOP", dopText, dopColor);
-        }
-        else if (statement.NonParallelPlanReason != null)
-            AddRow("Serial", statement.NonParallelPlanReason);
-
-        if (statement.QueryTimeStats != null)
-        {
-            AddRow("CPU", $"{statement.QueryTimeStats.CpuTimeMs:N0}ms");
-            if (statement.QueryUdfCpuTimeMs > 0)
-                AddRow("UDF CPU", $"{statement.QueryUdfCpuTimeMs:N0}ms");
-            if (statement.QueryUdfElapsedTimeMs > 0)
-                AddRow("UDF elapsed", $"{statement.QueryUdfElapsedTimeMs:N0}ms");
-        }
-
-        // Compile stats (category B plan-level property)
-        if (statement.CompileTimeMs > 0)
-            AddRow("Compile", $"{statement.CompileTimeMs:N0}ms");
-        if (statement.CachedPlanSizeKB > 0)
-            AddRow("Cached plan size", $"{statement.CachedPlanSizeKB:N0} KB");
-
-        // Memory grant — color per new tiers, spill indicator if any operator spilled
-        if (statement.MemoryGrant != null)
-        {
-            var mg = statement.MemoryGrant;
-            var grantPct = mg.GrantedMemoryKB > 0
-                ? (double)mg.MaxUsedMemoryKB / mg.GrantedMemoryKB * 100 : 100;
-            var grantColor = MemoryGrantColor(grantPct, hasSpillInTree);
-            var spillTag = hasSpillInTree ? " ⚠ spill" : "";
-            AddRow("Memory grant",
-                $"{TextFormatter.FormatMemoryGrantKB(mg.GrantedMemoryKB)} granted, {TextFormatter.FormatMemoryGrantKB(mg.MaxUsedMemoryKB)} used ({grantPct:N0}%){spillTag}",
-                grantColor);
-            if (mg.GrantWaitTimeMs > 0)
-                AddRow("Grant wait", $"{mg.GrantWaitTimeMs:N0}ms", "#E57373");
-        }
-
-        // Thread stats
-        if (statement.ThreadStats != null)
-        {
-            var ts = statement.ThreadStats;
-            AddRow("Branches", ts.Branches.ToString());
-            var totalReserved = ts.Reservations.Sum(r => r.ReservedThreads);
-            if (totalReserved > 0)
-            {
-                var threadPct = (double)ts.UsedThreads / totalReserved * 100;
-                var threadColor = EfficiencyColor(threadPct);
-                var threadText = ts.UsedThreads == totalReserved
-                    ? $"{ts.UsedThreads} used ({totalReserved} reserved)"
-                    : $"{ts.UsedThreads} used of {totalReserved} reserved ({totalReserved - ts.UsedThreads} inactive)";
-                AddRow("Threads", threadText, threadColor);
-            }
-            else
-            {
-                AddRow("Threads", $"{ts.UsedThreads} used");
-            }
-        }
-
-        // Optimization + CE model
-        if (!string.IsNullOrEmpty(statement.StatementOptmLevel))
-            AddRow("Optimization", statement.StatementOptmLevel);
-        if (!string.IsNullOrEmpty(statement.StatementOptmEarlyAbortReason))
-            AddRow("Early abort", statement.StatementOptmEarlyAbortReason);
-        if (statement.CardinalityEstimationModelVersion > 0)
-            AddRow("CE model", statement.CardinalityEstimationModelVersion.ToString());
-
-        if (grid.Children.Count > 0)
-        {
-            RuntimeSummaryContent.Children.Add(grid);
-            RuntimeSummaryEmpty.IsVisible = false;
-        }
-        else
-        {
-            RuntimeSummaryEmpty.IsVisible = true;
-        }
-        ShowServerContext();
-    }
-
-    private void ShowServerContext()
-    {
-        ServerContextContent.Children.Clear();
-        if (_serverMetadata == null)
-        {
-            ServerContextEmpty.IsVisible = true;
-            ServerContextBorder.IsVisible = true;
-            return;
-        }
-
-        ServerContextEmpty.IsVisible = false;
-
-        var m = _serverMetadata;
-        var fgColor = "#E4E6EB";
-
-        var grid = new Grid { ColumnDefinitions = new ColumnDefinitions("Auto,*") };
-        int rowIndex = 0;
-
-        void AddRow(string label, string value)
-        {
-            grid.RowDefinitions.Add(new RowDefinition(GridLength.Auto));
-            var lb = new TextBlock
-            {
-                Text = label, FontSize = 11,
-                Foreground = new SolidColorBrush(Color.Parse(fgColor)),
-                HorizontalAlignment = HorizontalAlignment.Left,
-                Margin = new Thickness(0, 1, 8, 1)
-            };
-            Grid.SetRow(lb, rowIndex);
-            Grid.SetColumn(lb, 0);
-            grid.Children.Add(lb);
-
-            var vb = new TextBlock
-            {
-                Text = value, FontSize = 11,
-                Foreground = new SolidColorBrush(Color.Parse(fgColor)),
-                Margin = new Thickness(0, 1, 0, 1)
-            };
-            Grid.SetRow(vb, rowIndex);
-            Grid.SetColumn(vb, 1);
-            grid.Children.Add(vb);
-            rowIndex++;
-        }
-
-        // Server name + edition
-        var edition = m.Edition;
-        if (edition != null)
-        {
-            var idx = edition.IndexOf(" (64-bit)");
-            if (idx > 0) edition = edition[..idx];
-        }
-        var serverLine = m.ServerName ?? "Unknown";
-        if (edition != null) serverLine += $" ({edition})";
-        if (m.ProductVersion != null) serverLine += $", {m.ProductVersion}";
-        AddRow("Server", serverLine);
-
-        // Hardware
-        if (m.CpuCount > 0)
-            AddRow("Hardware", $"{m.CpuCount} CPUs, {m.PhysicalMemoryMB:N0} MB RAM");
-
-        // Instance settings
-        AddRow("MAXDOP", m.MaxDop.ToString());
-        AddRow("Cost threshold", m.CostThresholdForParallelism.ToString());
-        AddRow("Max memory", $"{m.MaxServerMemoryMB:N0} MB");
-
-        // Database
-        if (m.Database != null)
-            AddRow("Database", $"{m.Database.Name} (compat {m.Database.CompatibilityLevel})");
-
-        ServerContextContent.Children.Add(grid);
-        ServerContextBorder.IsVisible = true;
-    }
-
     private void UpdateInsightsHeader()
     {
         InsightsPanel.IsVisible = true;
         InsightsHeader.Text = "  Plan Insights";
     }
 
-    private static string GetWaitCategory(string waitType)
-    {
-        if (waitType.StartsWith("SOS_SCHEDULER_YIELD") ||
-            waitType.StartsWith("CXPACKET") ||
-            waitType.StartsWith("CXCONSUMER") ||
-            waitType.StartsWith("CXSYNC_PORT") ||
-            waitType.StartsWith("CXSYNC_CONSUMER"))
-            return "CPU";
-
-        if (waitType.StartsWith("PAGEIOLATCH") ||
-            waitType.StartsWith("WRITELOG") ||
-            waitType.StartsWith("IO_COMPLETION") ||
-            waitType.StartsWith("ASYNC_IO_COMPLETION"))
-            return "I/O";
-
-        if (waitType.StartsWith("LCK_M_"))
-            return "Lock";
-
-        if (waitType == "RESOURCE_SEMAPHORE" || waitType == "CMEMTHREAD")
-            return "Memory";
-
-        if (waitType == "ASYNC_NETWORK_IO")
-            return "Network";
-
-        return "Other";
-    }
-
-    private static string GetWaitCategoryColor(string category)
-    {
-        return category switch
-        {
-            "CPU" => "#4FA3FF",
-            "I/O" => "#FFB347",
-            "Lock" => "#E57373",
-            "Memory" => "#9B59B6",
-            "Network" => "#2ECC71",
-            _ => "#6BB5FF"
-        };
-    }
 }
